gptel: Activate tool use
* gptel.el (gptel-request, gptel-use-tools, gptel-tools): Add user
options `gptel-tools' and `gptel-use-tools' to hold tools and as a
boolean to toggle tool use, respectively.  `gptel-tools' is meant
to hold a list of `gptel-tool' structs.

Presently `gptel-use-tools' is an all or nothing setting, adding
all defined tools to every request if it's enabled.  This setup is
for testing only, until we can find more flexible ways to specify
tools with requests.

* gptel-openai.el (gptel--request-data): If tools are enabled and
available, add them to the payload.

* gptel-ollama.el (gptel--request-data): Ditto.  Additionally,
disable streaming if using tools, since Ollama does
not (uniformly) support tool-use with streaming responses.

* gptel-gemini.el (gptel--request-data): Ditto.

* gptel-anthropic.el (gptel--request-data): Ditto.