gptel: Add a registry for tools
* gptel.el (gptel-tools, gptel-make-tool, gptel--known-tools):
`gptel-make-tool' now registers the tool in `gptel--known-tools'.
This is useful for updating a tools menu in gptel's transient UI,
which is a forthcoming change.  `gptel--known-tools' groups tools
by category and name.

Add three more slots to the `gptel-tool' struct: CATEGORY, CONFIRM
and INCLUDE.
- CATEGORY will be used to group tools in gptel's menus,
- CONFIRM determines if the tool call requires user confirmation.
- INCLUDE determines if tool results will be added to
chat buffers.  These names are tentative and subject to change.