gptel: Run callback with pending tool calls
* gptel.el (gptel-request, gptel--handle-tool-use): When
`gptel-confirm-tool-calls' is non-nil and/or a tool to be run
requires confirmation, run the `gptel-request' callback with a
list of pending tool calls.  Each element in this list is an alist
mapping the tool to be run (of type `gptel-tool') to the tool
handler callback and a list of the function arguments:

((tool1 tool-handler-cb1 arg-list1)
 (tool2 tool-handler-cb2 arg-list2)
 ...)

This differs from running the callback with the tool results,
where it is called with an alist of the form

((name1 . result1)
 (name1 . result2)
 (name2 . result3)
 ...)

(Note that the same tool may be called multiple times with
different results.)

The two kinds of calls can be differentiated via the type of the
alist key.

Fix typo in the documentation of `gptel-request'.

Neither pending calls nor tool results are handled yet by gptel's
default `gptel-request' callbacks.  This will be added next.